[py] Moved logging API to Chromium browsers only
